post

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->get();
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }
+            return view('welcome');
+
     }
 
     /**
@@ -33,7 +41,7 @@ class TasksController extends Controller
         
         return view('tasks.create', [
             'task' => $task,
-            ]);
+        ]);
     }
 
     /**
@@ -49,14 +57,13 @@ class TasksController extends Controller
             'content' => 'required|max:191',
             ]);
             
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->user_id = \Auth::id();
-        $task->save();
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
         return redirect('/');
-    }
+        }
 
     /**
      * Display the specified resource.
@@ -118,9 +125,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-        
+        $task = \App\Task::find($id);
+
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
         return redirect('/');
-    }
+        }
 }
